refactor(utils): extract terminal helpers into mtui/term.py
Move termsize, filter_ansi, prompt_user, and page verbatim out of
mtui/utils.py into a new mtui/term.py module. The four helpers stay
co-located because page() drives prompt_user() / termsize() /
filter_ansi() and the intra-module call sites are easier to follow
without crossing a module boundary.

utils.py keeps a re-export shim for callers so this commit is
byte-identical at the public-API surface.

Fourth commit of the Phase 5b / C8 split (PLAN.md item 11).
